fix: application startup for different databases
Moved the SQL script execution crosscheck from run() to executeSqlScript() in AugustApplication.java. Added database-specific scripts for h2 and PostgreSQL inside the method to give the suitable execution path. Updated the 'createAclSchema.sql' and 'createAclSchemaPostgres.sql' to ensure tables are not re-created if they already exist in the database, thus improving startup time.